refactor(stride_cli): Move argument parsing to clap
I would like to add option to select between "RepositoryTypes", and
as such needed option parsing.  If the decision not to use clap was a
deliberate one, and not a result of the early stages of development, then
I'm perfectly happy to implement that parsing by hand.  Nevertheless,
using clap brings further improvements, like an always up-to-date help
output, which I quite like.