feat(stride_flutter_bridge): Support `taskchampion` replicas
Currently, this "support" is in a very WIP state, but the general
API and the core though should become apparent.  I'm also not so sure
about the user experience with these changes, as both repositories are
clearly separated, forcing the user to `stride export` and the
`stride --repository task-champion import` the tasks again.