build(devenv): Provide frb to enable dart binding generation
I'm not so sure what the correct way to invoke frb is, so I simply tried
to find one that minimizes the resulting diff. This way is now
documented in `./frb_generate.sh` to make it easier for the next one
that searches for a way to regenerate the bindings.