fix(crates/flutter_bridge/src/api/settings.rs): Conditionally change $HOME variable
Changing $HOME leads to `dirs::*` returning their value relative to the
changed home (i.e., `~/.local/share/org.stridetasks.stride/.config`
instead of `~/.config`) in the absence of a `$XDG_*_HOME` variable.

This is necessary for android but rather confusing on a desktop system.
As such, make the change conditional to android systems.